prepare new code structure for VaultClient() class
